Work on handling api requests
Using an interactor model with response and request objects.
No specific boundary objects on the presenattion side are
required since this task is already handled by the EvenEmitter
system. Boundary objects on the data access side will be needed
later on, though it is not clear yet how these dependencies
can be managed well.