Implemented "reference" TypeSpec since "instance" alone is not enough conceptually
When using instance.of( SomeConstructor ) for a ModelDesign we can not use SomeConstructor to refer
to another ModelDesign's model implementation since the implementation has not been created yet.
Therefore we need another type which can be replaced with an instance or with a field for an ID
depending on the responsible model impelementer (also see new ModelImplementer.js file).